ml_training/requirements_ml.txt
# Core ML
tensorflow==2.15.0
keras==2.15.0

# Procesamiento de imágenes
Pillow==10.1.0
opencv-python==4.8.0.76

# Científicas
numpy==1.26.2
pandas==2.1.4

# Machine Learning
scikit-learn==1.3.2

# Visualización
matplotlib==3.8.2
seaborn==0.13.0

# Progress bars
tqdm==4.66.1

# Utilidades
python-dotenv==1.0.0